% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLandscape.R
\name{createLandscape}
\alias{createLandscape}
\title{Landscape function}
\usage{
createLandscape(r, size_pp, n_nf, size_nf)
}
\arguments{
\item{r}{A raster object}

\item{size_pp}{The size of pine plantations expressed as number of pixels (integer)}

\item{n_nf}{Number of patchs of Natural forests (a number between 1 and 5)}

\item{size_nf}{The size of natural forest patchs (a value between 10 and 500)}
}
\value{
A raster object with the landscape confifured
}
\description{
Landscape function
}
\examples{
# m <- matrix(nrow=100, ncol=200, byrow = TRUE)
# r <- raster(m)
# extent(r) <- matrix(c(0, 0, 200, 100), nrow=2)
# r[] <- 0
# myl <- createLandscape(r, size_pp = 1000, size_nf = 500, n_nf = 4)
}
