% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2nf.R
\name{dist2nf}
\alias{dist2nf}
\title{Distance to Natural forests}
\usage{
dist2nf(r, nf_value)
}
\arguments{
\item{r}{A \code{raster} object}

\item{nf_value}{The value of "Natural Forests" class within the raster (Default value = 2)}
}
\value{
A \code{raster} object with the minimun distance for each raster cell
}
\description{
Compute the distance between each pixel and the edges of all natural forests.
Then get the minimun distance for each pixel.
}
\author{
Antonio J Perez-Luque
}
