% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initRichness.R
\name{initRichness}
\alias{initRichness}
\title{Set values of Initial Richness}
\usage{
initRichness(r, r_range, treedensity, pastUse, rescale = TRUE)
}
\arguments{
\item{r}{A \code{raster} object}

\item{r_range}{A \code{data frame} with three columns: \code{value} of land use
(\code{integer}: 0 = "Other", 1 = "Pine plantation", 2 = "Natural Forests",
3 = "Crop"); \code{lowRich} and \code{upRich} (lower an upper value of the
range of Richness: See Gomez-Aparicio et al 2009)}

\item{treedensity}{density of the pine plantation (\code{integer})}

\item{pastUse}{the past land use of the pine plantation (\code{character}).
One of "Oak", "Shrubland", "Pasture" or "Crop"}

\item{rescale}{If "TRUE" the results are rescaled}
}
\value{
A \code{raster} object with values of initial Richness for each
pixel.
}
\description{
Esta funcion establece los valores iniciales de riqueza de especies
}
\references{
Gomez-Aparicio L, Zavala MA, Bonet FJ, Zamora R (2009) Are pine plantations
valid tools for restoring Mediterranean forests? An assessment along abiotic
and biotic gradients. Ecological Applications, 19: 2124–2141.
}
