% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{richPot}
\alias{richPot}
\title{Potential Richness of Sierra Nevada forests}
\format{A \code{data frame} with three rows and 5 columns:
\describe{
 \item{eco}{Ecosystem type}
 \item{n}{number of plots used to estimate the potential richness. See methods
   section in Gomez-Aparicio et al. (2009)}
 \item{potRich}{predicted values of Richness computed using Eq. 1 of
   Gomez-Aparicio et al. (2009)}
 \item{lowerInterval}{lowere value of the Support interval}
 \item{upperInterval}{lowere value of the Support interval}
}}
\usage{
richPot
}
\description{
A dataset containing the potential richness of Pine plantantios, natural
\emph{Quercus ilex} forests and deciduous forests. The potential richness
computed using the equations provided by Gomez-Aparicio et al. (2009). See
Equation 1 and Appendix A2 of this paper.
}
\details{
Gomez-Aparicio L, Zavala MA, Bonet FJ, Zamora R (2009) Are pine plantations
valid tools for restoring Mediterranean forests? An assessment along abiotic
and biotic gradients. Ecological Applications, 19: 2124 - 2141.
}
\references{
\url{https://dx.doi.org/10.1890/08-1656.1}
}
\keyword{datasets}
