% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{richSinfo_agg}
\alias{richSinfo_agg}
\title{Richness of Sierra Nevada (Spain) forests from SINFONEVADA aggregated}
\format{A \code{data frame} with 11 rows and 9 columns:
\describe{
 \item{newECO}{name of the ecosystem type}
 \item{mean}{richnnes mean value of the plots within this ecosystem type}
 \item{sd}{standard deviation of the richness mean value of the plots within
  this ecosystem type}
 \item{se}{standard error of the richness mean value of the plots within
  this ecosystem type}
 \item{n}{number of plots within this ecosystem type}
 \item{min}{minimun value of richnnes of the plots within this ecosystem type}
 \item{max}{maximum value of richnnes of the plots within this ecosystem type}
 \item{median}{richnnes median value of the plots within this ecosystem type}
}}
\usage{
richSinfo_agg
}
\description{
A dataset containing the richness of Sierra Nevada ecosystems according to
Sinfonevada (Perez-Luque et al. 2014). The richness value of each plot
was used to compute mean values by ecosystem types using a spatial overlay of
the plots and the ecosystems map of Sierra Nevada. Then the ecoystem were
reclassify.
}
\details{
Pérez-Luque A, Bonet F, Pérez-Pérez R, Aspizua R, Lorite J, Zamora R (2014)
Sinfonevada: Dataset of Floristic diversity in Sierra Nevada
forests (SE Spain). PhytoKeys 35: 1-15.
}
\references{
\url{https://doi.org/10.3897/phytokeys.35.6363}
}
\keyword{datasets}
